docs/references/reference-moha-bot-v1/requirements.txt
# Flask API
flask==3.0.0
flask-cors==4.0.0
gunicorn==21.2.0

# Blockchain - let pip resolve compatible versions
eth-account>=0.9.0
hyperliquid-python-sdk

# Utilities
python-dotenv==1.0.0
requests==2.31.0
pydantic-settings>=2.0.0
